backend/requirements.txt
fastapi
uvicorn[standard]
openai
faiss-cpu
numpy
redis
python-dotenv
pydantic
sqlalchemy
python-jose[cryptography]
python-multipart
google-auth
google-auth-oauthlib
requests
pypdf
python-docx
Pillow
reportlab
psycopg2-binary
stripe>=7.0.0
